requirements.txt
fastmcp
dotenv
httpx
